app/requirements.txt
asgiref==3.7.2
backcall==0.2.0
blessings==1.7
boto3==1.34.11
botocore==1.34.11
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
click==8.1.7
coverage==7.4.0
cryptography==41.0.7
datapunt-keycloak-oidc @ git+https://github.com/remyvdwereld/keycloak_oidc.git
debtcollector==2.5.0
decorator==5.1.1
dj-database-url==2.1.0
Django==5.0
django-constance==3.1.0
django-cors-headers==4.3.1
django-filter==23.5
django-formtools==2.5.1
django-model-utils==4.3.1
django-multiselectfield @ git+https://github.com/remyvdwereld/django-multiselectfield.git # Latest version of django-multiselectfield (0.1.12) not compatible with Django=5 so use this fork.
django-nose==1.4.7
django-picklefield==3.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-storage-swift==1.3.0
# NEXT LINES TODO
django-storages==1.12.3
django-unique-upload==0.2.1
django-webpack-loader==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.17.0
entrypoints==0.3
factory-boy==2.11.1
Faker==2.0.4
flake8==3.7.5
html5lib==1.1
idna==2.8
importlib-metadata==4.11.3
inflection==0.5.1
ipython==8.19.0
iso8601==1.0.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.0
josepy==1.13.0
keystoneauth1==4.5.0
livereload==2.6.3
Markdown==3.3.6
MarkupSafe==2.1.1
mccabe==0.6.1
mkdocs==1.5.3
mock==2.0.0
mozilla-django-oidc==3.0.0
msgpack==1.0.3
netaddr==0.8.0
netifaces==0.11.0
newrelic==4.12.0.113
nose==1.3.7
os-service-types==1.7.0
oslo.config==8.8.0
oslo.i18n==5.1.0
oslo.serialization==4.3.0
oslo.utils==4.12.2
packaging==21.3
parso==0.8.3
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.43
psycopg2==2.9.3
ptyprocess==0.7.0
pycodestyle==2.5.0
pycparser==2.21
pyflakes==2.1.1
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-http-client==3.3.7
python-keystoneclient==4.4.0
python-magic==0.4.25
python-swiftclient==3.13.1
pytz==2023.3
PyYAML==6.0.1
regex==2022.3.15
requests==2.31.0
rfc3986==2.0.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
sendgrid==6.2.0
sentry-sdk==1.39.1
six==1.16.0
sqlparse==0.4.4
stevedore==3.5.0
text-unidecode==1.3
textile==4.0.1
tornado==6.4
traitlets==5.1.1
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.0
zipp==3.7.0
